fix: resolved all code changes requested
