+++ apps/demo-identity-app/package.json
@@ -11,6 +11,7 @@
     "format": "prettier --write ."
   },
   "dependencies": {
+    "@farcaster/miniapp-sdk": "^0.1.8",
     "@goodsdks/citizen-sdk": "*",
     "@goodsdks/react-hooks": "*",
     "@reown/appkit": "^1.7.2",
@@ -34,7 +35,6 @@
     "zod": "^3.24.2"
   },
   "devDependencies": {
-    "@farcaster/miniapp-sdk": "^0.1.8",
     "@tamagui/babel-plugin": "^1.125.22",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
