+++ apps/demo-webcomponents/vite.config.mts
@@ -26,9 +26,5 @@ export default defineConfig({
   define: {
     "process.env": process.env,
   },
-  build: {
-    rollupOptions: {
-      external: ["@goodsdks/citizen-sdk", "viem"]
-    }
-  }
+
 })
