+++ packages/citizen-sdk/src/index.ts
@@ -1,4 +1,4 @@
 export * from "./sdks"
 export * from "./constants"
 export * from "./utils/auth"
-export * from "./utils/farcaster-navigation"
+
